+++ src/scheduler/scheduler.service.ts
@@ -13,9 +13,9 @@ export class SchedulerService {
   constructor(@InjectQueue(SCHEDULER_QUEUE) private queue: Queue) {}
 
   async addJob<T>(jobName: string, data?: T, delay = 0, jobId?: string) {
-    if (process.env.NODE_ENV === 'development') {
-      return;
-    }
+    // if (process.env.NODE_ENV === 'development') {
+    //   return;
+    // }
 
     try {
       if (!!jobId) {